+++ libdimension/dimension/gl.h
@@ -33,7 +33,7 @@ int dmnsn_gl_optimize_canvas(dmnsn_canvas *canvas);
 int dmnsn_gl_write_canvas(const dmnsn_canvas *canvas);
 
 /* Read a canvas from a GL framebuffer.  Returns NULL on failure. */
-dmnsn_canvas *dmnsn_gl_read_canvas(unsigned int x0, unsigned int y0,
-                                   unsigned int width, unsigned int height);
+dmnsn_canvas *dmnsn_gl_read_canvas(size_t x0, size_t y0,
+                                   size_t width, size_t height);
 
 #endif /* DIMENSION_GL_H */
